requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
simpy==4.1.1
Werkzeug==3.1.3